requirements.txt
discord.py==1.7.2
tweepy==3.10.0
